ready el rating
