+++ Client/urLearning/src/components/AddComment/AddComment.jsx
@@ -3,7 +3,12 @@ import { Label, Textarea, Button } from "flowbite-react"
 import { useState } from 'react'
 import { useParams } from "react-router-dom";
 import { useDispatch } from 'react-redux';
-import { getComment, postComment } from '../../redux/actions';
+import { getComment, postComment, putRating } from '../../redux/actions';
+import style from "../AddComment/AddComment.module.css"
+import corazon from "../../img/IoIosHeart.svg"
+import { FaStar } from "react-icons/fa";
+
+// import AddRating from '../AddRating/AddRating';
 
 
 function AddComment() {
@@ -13,6 +18,11 @@ function AddComment() {
 
     const { id } = useParams()
     const dispatch = useDispatch()
+    const [currentValue, setcurrentValue] = useState({
+        rating: ""
+    })
+    const [hoverValue, sethoverValue] = useState(undefined)
+    const stars = Array(5).fill(0)
 
 
     const handleChange = (ev) => {
@@ -22,22 +32,55 @@ function AddComment() {
         });
     }
 
+    const handleClick = (value) => {
+        setcurrentValue(value)
+
+    }
+
+    const handleMouse = (newValue) => {
+        sethoverValue(newValue)
+    }
+
+    const handleMouseLeave = () => {
+        sethoverValue(undefined)
+    }
 
     const handleSubmit = (ev) => {
         ev.preventDefault();
         dispatch(postComment(id, input));
+        dispatch(putRating(id, currentValue))
         setInput({
             comment: ""
         });
         dispatch(getComment())
+        sethoverValue(undefined)
+        setcurrentValue({
+            rating: ""
+        })
 
     };
+
+    console.log(currentValue)
 
 
     return (
         <div>
             <form onSubmit={(e) => handleSubmit(e)}>
-
+                <div className={style.estrellas}>
+                    {stars.map((_, index) => {
+                        return (
+                            <div>
+                                <FaStar
+                                    className={(hoverValue || currentValue) > index ? style.amarillo : style.gris}
+                                    key={index}
+                                    onClick={() => handleClick(index + 1)}
+                                    onMouseOver={() => handleMouse(index + 1)}
+                                    onMouseLeave={handleMouseLeave}
+                                />
+                            </div>
+                        )
+                    })}
+                </div>
                 <div id="textarea">
                     <div className="mb-2 block">
                         <Label
@@ -57,6 +100,7 @@ function AddComment() {
                 <Button
                     gradientDuoTone="purpleToBlue"
                     type="submit"
+                    className={style.boton}
                 >
                     Comentar
                 </Button>
