--- Client/urLearning/src/components/AddRating/AddRating.module.css
@@ -1,7 +1,0 @@
-.gris {
-  filter: grayscale(0.9);
-}
-
-.estrellas {
-  display: flex;
-}
